Update README.md
Hey, I came across this packages after searching the internet for a good directive to use with Disqus, and I think this is the best one so far. I hope to make it even better by first editing the README.md.

I made some edits to the 'Getting Started' section by adding steps. I also added code examples for the steps. I kept only the provider method for setting the shortname.

It took me extra time to understand the doc because the lack of concrete examples. I hope to made this package even better by adding more examples when appropriate.

- Webber